Add general luxld linker.
This linker is the result of the combination of the existing linkers
plus a new linker for ARM. Also, MIPS shared libraries are supported now.

The driver was extensively modified to support shared libraries and also
to make use of the new linker.

The assemblers now support a 'function' extension for the 'global' directive.

musl support was improved.

+++ src/lib/include/stdarg.h
@@ -41,7 +41,7 @@ static void *__va_arg(va_list ap, unsigned long siz)
     void *p;
     int num_gp;
 
-    num_gp = _LONGSIZE(siz)/8;
+    num_gp = (int)(_LONGSIZE(siz)/8);
 
     if (siz>16 || ap->gp_offset>48-num_gp*8) {
         p = ap->overflow_arg_area;
